requirements.txt
certifi==2021.5.30
expiringdict==1.1.4
launchdarkly-server-sdk==7.2.0
pyRFC3339==1.1
python-dotenv==0.18.0
pytz==2021.1
semver==2.13.0
urllib3==1.26.6
